fix: benchmark.filter_options
To allow for other intra-bilan categories (scope, emission category,
...).

For some reason, the line

```
x = x[x != 0]
```

was not dropping any rows, instead it was replacing all the 0. with nan
(at least in the emission columns).

The end goal is to exclude nan and 0. from the computation when plotting
emissions (mean, median). By chance, this was working (since nans are
dropped when plotting the barplot), but it was a bit by mistake.

To make it clearer, we now simply drop the rows in question.

Also, in this commit:
* adapt the tests to this change
* adapt aggregate_bilans to other intra-bilan categories
* add new filtering options